Stupid Layout Spacing Script
Stupid implementation with layout spacing script.
If anyone ever reads this, the most annoying and difficult task about this project is to ensure that spacing between all elements are a percentage of the screens size. So that it can work exactly the same on all resolutions.
It is astonishing how difficult this is to achieve. I've been using Unity and the Editor for quite a while, but never came across a satisfying solution to this. As I want to make a perfect project because the position means a lot to me, i just spent probably more than 12 hours tinkering with the layouts just to make sure the spacing is not off by 0.0125 in some extreme screen sizes. But, I guess, the job demands perfection. After all, the design layout you've sent me has the spacings exact. I was hoping to be done with UI 10 hours ago and started implementing callbacks, localization and themes functionality. Here I am after a whole weekend waster on UI Spacing and have only disappointment in me. I've done an hour worth of work in 12 hours. Just to be perfect to the tiniest millimeter. And even still, I don't even know if this is actually required and will be "graded".